+++ packages/iris-grid/src/sidebar/visibility-ordering-builder/sortable-tree/TreeItem.tsx
@@ -1,13 +1,14 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import React, { useMemo } from 'react';
 import classNames from 'classnames';
-import type { FlattenedItem, TreeItem as TreeItemType } from './types';
+import type { FlattenedItem } from './types';
 import './TreeItem.scss';
 
-export interface Props<T> {
+export interface TreeItemProps<T> {
   childCount?: number;
   clone?: boolean;
   depth: number;
+  withDepthMarkers?: boolean;
   disableInteraction?: boolean;
   ghost?: boolean;
   handleProps?: Record<string, unknown>;
@@ -15,29 +16,31 @@ export interface Props<T> {
   item: FlattenedItem<T>;
   dragRef?: React.Ref<HTMLDivElement> | null;
   wrapperRef?: React.Ref<HTMLLIElement> | null;
-  renderItem: (props: {
-    ref: React.Ref<HTMLDivElement> | null;
-    clone: boolean;
-    childCount?: number;
-    value: string;
-    item: FlattenedItem<T>;
-    handleProps?: Record<string, unknown>;
-  }) => JSX.Element;
+  renderItem: TreeItemRenderFn<T>;
+  /**
+   * Styles from dnd-kit to transform the depth lines
+   */
+  style?: React.CSSProperties;
 }
 
-export type TreeItemRenderFn<T> = (props: {
+export type TreeItemRenderFnProps<T> = {
   ref: React.Ref<HTMLDivElement> | null;
   clone: boolean;
   childCount?: number;
   value: string;
-  item: T extends TreeItemType<infer D> ? FlattenedItem<D> : FlattenedItem<T>;
+  item: FlattenedItem<T>;
   handleProps?: Record<string, unknown>;
-}) => JSX.Element;
+};
 
-export function TreeItem<T>(props: Props<T>): JSX.Element {
+export type TreeItemRenderFn<T> = (
+  props: TreeItemRenderFnProps<T>
+) => JSX.Element;
+
+export function TreeItem<T>(props: TreeItemProps<T>): JSX.Element {
   const {
     clone = false,
     depth,
+    withDepthMarkers = true,
     disableInteraction = false,
     ghost = false,
     handleProps,
@@ -47,15 +50,24 @@ export function TreeItem<T>(props: Props<T>): JSX.Element {
     renderItem,
     item,
     childCount,
+    style,
   } = props;
 
   const depthMarkers = useMemo(
     () =>
-      Array(depth)
-        .fill(0)
-        // eslint-disable-next-line react/no-array-index-key
-        .map((_, i) => <span key={`depth-line-${i}`} className="depth-line" />),
-    [depth]
+      withDepthMarkers
+        ? Array(depth)
+            .fill(0)
+            .map((_, i) => (
+              <span
+                // eslint-disable-next-line react/no-array-index-key
+                key={`depth-line-${i}`}
+                className="depth-line"
+                style={style}
+              />
+            ))
+        : null,
+    [depth, style, withDepthMarkers]
   );
 
   const renderItemProps = useMemo(
@@ -78,9 +90,10 @@ export function TreeItem<T>(props: Props<T>): JSX.Element {
         ghost,
         disableInteraction,
       })}
+      data-id={item.id}
       ref={wrapperRef}
     >
-      {!clone && depthMarkers}
+      {!clone && withDepthMarkers && depthMarkers}
       {renderItem(renderItemProps)}
     </li>
   );